Extract Favorite to FavoriteEntity mapping
And fix a few cases with price alert input.

+++ data/core/src/main/kotlin/saschpe/gameon/data/core/model/Favorite.kt
@@ -4,5 +4,5 @@ data class Favorite(
     val id: Int? = null,
     val plain: String,
     val createdAt: Long = System.currentTimeMillis(),
-    val priceThreshold: Long? = null
+    val priceThreshold: Double? = null
 )
